Fixed some issues with the connection between the CLI and the Service

+++ final/ticket-service/src/main/java/com/epam/training/ticketservice/repository/impl/JpaMovieRepository.java
@@ -6,6 +6,7 @@ import com.epam.training.ticketservice.domain.Movie;
 import com.epam.training.ticketservice.repository.MovieRepository;
 import org.springframework.stereotype.Repository;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 
@@ -19,13 +20,18 @@ public class JpaMovieRepository implements MovieRepository {
     }
 
     @Override
-    public void createMovie(Movie movie) {
-        movieDao.save(new MovieProjection(
-                null,
-                movie.getMovieTitle(),
-                movie.getMovieGenre(),
-                movie.getMovieLength()
-        ));
+    public void createMovie(Movie movie) throws Exception {
+        if (movieDao.findByMovieTitle(movie.getMovieTitle()).isEmpty()){
+            movieDao.save(new MovieProjection(
+                    null,
+                    movie.getMovieTitle(),
+                    movie.getMovieGenre(),
+                    movie.getMovieLength()
+            ));
+        }
+        else {
+            throw new Exception("Movie with title '" + movie.getMovieTitle() + "' is already exists!");
+        }
     }
 
     @Override
@@ -39,9 +45,9 @@ public class JpaMovieRepository implements MovieRepository {
     }
 
     @Override
-    public Movie findMovieByTitle(String movieTitle) throws Exception {
+    public Movie findMovieByMovieTitle(String movieTitle) throws Exception {
         MovieProjection movieProjection = movieDao.findByMovieTitle(movieTitle).orElseThrow(
-                () -> new Exception("Movie not found with \""+ movieTitle +"\" title!")
+                () -> new Exception("Movie not found with '"+ movieTitle +"' title!")
         );
 
         return new Movie(
@@ -51,9 +57,9 @@ public class JpaMovieRepository implements MovieRepository {
     }
 
     @Override
-    public void updateMovie(Movie movie) throws Exception {
+    public void updateMovieByMovie(Movie movie) throws Exception {
         MovieProjection movieProjection = movieDao.findByMovieTitle(movie.getMovieTitle()).orElseThrow(
-                () -> new Exception("Movie not found with \""+ movie.getMovieTitle() +"\" title!")
+                () -> new Exception("Movie not found with '"+ movie.getMovieTitle() +"' title!")
         );
         movieProjection.setMovieGenre(movie.getMovieGenre());
         movieProjection.setMovieLength(movie.getMovieLength());
@@ -61,7 +67,13 @@ public class JpaMovieRepository implements MovieRepository {
     }
 
     @Override
-    public void deleteMovie(String movieTitle) {
-        movieDao.deleteByMovieTitle(movieTitle);
+    @Transactional
+    public void deleteMovieByMovieTitle(String movieTitle) throws Exception {
+        if (movieDao.findByMovieTitle(movieTitle).isEmpty()) {
+            throw new Exception("Movie not found with '"+ movieTitle +"' title!");
+        }
+        else {
+            movieDao.deleteByMovieTitle(movieTitle);
+        }
     }
 }
